update to 3.0.0

+++ tests/app.test.js
@@ -9,10 +9,14 @@ beforeAll(async (done) => {
 
 /** this code is called once before all the tested are finished */
 afterAll(async (done) => {
+  const dbSettings = strapi.config.get("database.connections.default.settings");
+
   //delete test database after all tests
-  const tmpDbFile = `${__dirname}/../${strapi.config.currentEnvironment.database.connections.default.settings.filename}`;
-  if (fs.existsSync(tmpDbFile)) {
-    fs.unlinkSync(tmpDbFile);
+  if (dbSettings && dbSettings.filename) {
+    const tmpDbFile = `${__dirname}/../${dbSettings.filename}`;
+    if (fs.existsSync(tmpDbFile)) {
+      fs.unlinkSync(tmpDbFile);
+    }
   }
   done();
 });
